requirements.txt
pandas
numpy
matplotlib
h5py
boto3
fastapi
uvicorn
requests
pydantic
gunicorn
streamlit
geopy
urllib3
typing
imageio
tensorflow
python-jose[cryptography]
passlib[bcrypt]
python-multipart
simplejson
google-cloud-bigquery
google-cloud
